DepositionGeant4: cast where cast belongs

+++ src/modules/DepositionGeant4/SensitiveDetectorActionG4.cpp
@@ -75,9 +75,9 @@ G4bool SensitiveDetectorActionG4::ProcessHits(G4Step* step, G4TouchableHistory*)
 
     // Calculate number of electron hole pairs produced, taking into account fluctuations between ionization and lattice
     // excitations via the Fano factor. We assume Gaussian statistics here.
-    auto mean_charge = static_cast<unsigned int>(edep / charge_creation_energy_);
+    auto mean_charge = edep / charge_creation_energy_;
     std::normal_distribution<double> charge_fluctuation(mean_charge, std::sqrt(mean_charge * fano_factor_));
-    auto charge = charge_fluctuation(random_generator_);
+    auto charge = static_cast<unsigned int>(charge_fluctuation(random_generator_));
 
     auto deposit_position_g4loc =
         ROOT::Math::XYZPoint(deposit_position_g4.x() + detector_->getModel()->getSensorCenter().x(),
